+++ Sekoia.io/CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Handle empty `version` in `metadata`.
 
+## 2025-06-03 - 2.68.4
+
+### Fixed
+
+- Add retry logic when using GetAggregationQuery for ws connection.
+
 ## 2025-06-03 - 2.68.3
 
 ### Fixed